// MECProject-App/src/app/models/userCompleteDetails.model.ts
export class userCompleteDetails {
    user_ID: number
    user_name: string;
    user_email: string;
    user_phone: string;
    user_password: string;
    userstatus: string;
    shopkeeper_fname: string;
    shopkeeper_shopname: string;
    shopkeeper_shopaddress: string;
    shopkeeper_bank_name: string;
    shopkeeper_bank_branch: string;
    shopkeeper_bank_account: string;
    shopkeeper_commission: string;
    role_ID: string;
}